requirements.txt
python-telegram-bot
mongoengine
dnspython
